dsge_latvia/requirements.txt
numpy>=1.26
scipy>=1.11
pandas>=2.1
pyyaml>=6.0
sympy>=1.12
